+++ application/config/routes.php
@@ -51,5 +51,6 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 */
 $route['default_controller'] = 'welcome';
 $route['sleep'] = 'first/zzz';
+$route['show/(:num)'] = 'first/gimme/$1';
 $route['404_override'] = '';
 $route['translate_uri_dashes'] = FALSE;