+++ application/controllers/Welcome.php
@@ -29,5 +29,18 @@ class Welcome extends Application
 
 		$this->render();
 	}
+	
+	public function shucks()
+	{
+		// this is the view we want shown
+		$this->data['pagebody'] = 'justone';
 
+		// build the list of authors, to pass on to our view
+		$source = $this->quotes->get(2);
+		$this->data['mug'] = $source['mug'];
+		$this->data['who'] = $source['who'];
+		$this->data['what'] = $source['what'];
+
+		$this->render();
+	}
 }